+++ src/modules/instruments/hsi.cc
@@ -166,7 +166,7 @@ HSI::estimate_track()
 		_positions[0] = current_position;
 	}
 
-	double len10 = _positions[1].haversine_nm (_positions[0]);
+	float len10 = _positions[1].haversine_nm (_positions[0]);
 	Angle alpha = -180.0_deg + great_arcs_angle (_positions[2], _positions[1], _positions[0]);
 	Angle beta_per_mile = alpha / len10;
